requirements.txt
Django==4.0.6
djangorestframework==3.13.1
python-dotenv==0.20.0
pymongo==4.3.3
mongoengine==0.24.2
celery==5.2.7
redis==4.5.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
